task-tracker/pyproject.toml
[tool.poetry]
name = "task-tracker"
version = "0.1.0"
description = "UberPopug task tracker"
authors = ["Nikita Vstovsky <nikvst@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
pydantic = {extras = ["email"], version = "^2.1.1"}
sqlalchemy = "^2.0.19"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
httpx = "^0.24.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqladmin = "^0.14.1"
alembic = "^1.11.2"
pydantic-settings = "^2.0.2"
psycopg2-binary = "^2.9.7"
asyncpg = "^0.28.0"
itsdangerous = "^2.1.2"
confluent-kafka = "^2.2.0"
aiokafka = "^0.8.1"
greenlet = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
black = "^23.7.0"
pytest = "^7.4.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]

[tool.mypy]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "app.*"
]
ignore_missing_imports = true
follow_imports = "normal"
disallow_untyped_defs = true
disallow_untyped_calls = true
